+++ src/test/kotlin/it/pagopa/ecommerce/payment/requests/repositories/redistemplate/v1/CartsRedisTemplateWrapperTest.kt
@@ -1,19 +1,24 @@
-package it.pagopa.ecommerce.payment.requests.repositories.redistemplate
+package it.pagopa.ecommerce.payment.requests.repositories.redistemplate.v1
 
 import it.pagopa.ecommerce.payment.requests.domain.RptId
-import it.pagopa.ecommerce.payment.requests.repositories.CartInfo
 import it.pagopa.ecommerce.payment.requests.repositories.PaymentInfo
-import it.pagopa.ecommerce.payment.requests.repositories.ReturnUrls
+import it.pagopa.ecommerce.payment.requests.repositories.v1.CartInfo
+import it.pagopa.ecommerce.payment.requests.repositories.v1.ReturnUrls
 import it.pagopa.ecommerce.payment.requests.tests.utils.v1.CartRequests
 import java.time.Duration
-import java.util.*
-import org.junit.jupiter.api.Assertions.assertEquals
+import java.util.UUID
+import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import org.mockito.ArgumentCaptor
 import org.mockito.Captor
 import org.mockito.junit.jupiter.MockitoExtension
-import org.mockito.kotlin.*
+import org.mockito.kotlin.capture
+import org.mockito.kotlin.eq
+import org.mockito.kotlin.given
+import org.mockito.kotlin.mock
+import org.mockito.kotlin.times
+import org.mockito.kotlin.verify
 import org.springframework.data.redis.core.ReactiveRedisTemplate
 import org.springframework.data.redis.core.ReactiveValueOperations
 import reactor.core.publisher.Mono
@@ -51,8 +56,7 @@ class CartsRedisTemplateWrapperTest {
             ReturnUrls(
               returnSuccessUrl = it.returnOkUrl.toString(),
               returnErrorUrl = it.returnErrorUrl.toString(),
-              returnCancelUrl = it.returnCancelUrl.toString(),
-              null)
+              returnCancelUrl = it.returnCancelUrl.toString())
           },
           req.emailNotice)
       }
@@ -65,7 +69,7 @@ class CartsRedisTemplateWrapperTest {
 
     verify(redisTemplate, times(1)).opsForValue()
     verify(valueOperations, times(1)).set("carts:$cartId", cartInfo, duration)
-    assertEquals("carts:$cartId", keyCaptor.value)
+    Assertions.assertEquals("carts:$cartId", keyCaptor.value)
   }
 
   @Test
@@ -86,8 +90,7 @@ class CartsRedisTemplateWrapperTest {
             ReturnUrls(
               returnSuccessUrl = it.returnOkUrl.toString(),
               returnErrorUrl = it.returnErrorUrl.toString(),
-              returnCancelUrl = it.returnCancelUrl.toString(),
-              null)
+              returnCancelUrl = it.returnCancelUrl.toString())
           },
           req.emailNotice)
       }
@@ -101,70 +104,5 @@ class CartsRedisTemplateWrapperTest {
       .verifyComplete()
 
     verify(redisTemplate, times(1)).opsForValue()
-  }
-
-  @Test
-  fun `Should save entity with waitingUrl successfully`() {
-    val cartId = UUID.randomUUID()
-    val cartRequest =
-      CartRequests.withMultiplePaymentNotices(3, "www.comune.di.prova.it/pagopa/waiting.html")
-
-    val cartInfo =
-      cartRequest.let { req ->
-        CartInfo(
-          cartId,
-          req.paymentNotices.map {
-            PaymentInfo(
-              RptId(it.fiscalCode + it.noticeNumber), it.description, it.amount, it.companyName)
-          },
-          req.idCart,
-          req.returnUrls.let {
-            ReturnUrls(
-              returnSuccessUrl = it.returnOkUrl.toString(),
-              returnErrorUrl = it.returnErrorUrl.toString(),
-              returnCancelUrl = it.returnCancelUrl.toString(),
-              returnWaitingUrl = it.returnWaitingUrl?.toString())
-          },
-          req.emailNotice)
-      }
-
-    given(redisTemplate.opsForValue()).willReturn(valueOperations)
-    given(valueOperations.set(any(), eq(cartInfo), eq(duration))).willReturn(Mono.just(true))
-
-    StepVerifier.create(cartsRedisTemplateWrapper.save(cartInfo)).expectNext(true).verifyComplete()
-  }
-
-  @Test
-  fun `Should get entity with waitingUrl successfully`() {
-    val cartId = UUID.randomUUID()
-
-    val cartRequest =
-      CartRequests.withMultiplePaymentNotices(3, "www.comune.di.prova.it/pagopa/waiting.html")
-
-    val cartInfo =
-      cartRequest.let { req ->
-        CartInfo(
-          cartId,
-          req.paymentNotices.map {
-            PaymentInfo(
-              RptId(it.fiscalCode + it.noticeNumber), it.description, it.amount, it.companyName)
-          },
-          req.idCart,
-          req.returnUrls.let {
-            ReturnUrls(
-              returnSuccessUrl = it.returnOkUrl.toString(),
-              returnErrorUrl = it.returnErrorUrl.toString(),
-              returnCancelUrl = it.returnCancelUrl.toString(),
-              returnWaitingUrl = it.returnWaitingUrl?.toString())
-          },
-          req.emailNotice)
-      }
-
-    given(redisTemplate.opsForValue()).willReturn(valueOperations)
-    given(valueOperations.get("carts:$cartId")).willReturn(Mono.just(cartInfo))
-
-    StepVerifier.create(cartsRedisTemplateWrapper.findById(cartId.toString()))
-      .expectNext(cartInfo)
-      .verifyComplete()
   }
 }
